refactor: 프로필 기본 SVG fallback 제거 및 UserT를 types 파일로 정리
