+++ apps/web/src/components/user-profile-group/userProfile.types.ts
@@ -1,0 +1,5 @@
+export type UserT = {
+  id: string;
+  name: string;
+  imageUrl: string;
+};
